Fixed issue [security] #19252: Bad permission check when edit quota Dev: always control if quota->sid are related to surveyid param Dev: add some POST action control

+++ application/controllers/admin/Quotas.php
@@ -58,13 +58,27 @@ class Quotas extends SurveyCommonAction
     }
 
     /**
+     * Check permission cfor a survey about quota
+     * @param integer $iSurveyId
      * @param string $sPermission
+     * @param integer|null $iQuotaid
+     * throw Exception
+     * @return void
      */
-    private function checkPermissions($iSurveyId, $sPermission)
+    private function checkPermissions($iSurveyId, $sPermission, $iQuotaid = null)
     {
         if (!empty($sPermission) && !(Permission::model()->hasSurveyPermission($iSurveyId, 'quotas', $sPermission))) {
             Yii::app()->session['flashmessage'] = gT('Access denied!');
             $this->redirectToIndex($iSurveyId);
+        }
+        if ($iQuotaid) {
+            $model = Quota::model()->find(
+                "id = :quotaid AND sid = :surveyid",
+                [':quotaid' => $iQuotaid, ':surveyid' => $iSurveyId]
+            );
+            if (empty($model)) {
+                throw new CHttpException(403, gT("Bad quota id with this survey id."));
+            }
         }
     }
 
@@ -91,11 +105,23 @@ class Quotas extends SurveyCommonAction
                 /** @var Quota $oQuota */
                 $oQuota = Quota::model()->findByPk($iQuotaId);
                 if (in_array($action, array('activate', 'deactivate'))) {
+                    if (!Permission::model()->hasSurveyPermission($oQuota->sid, 'quotas', 'update')) {
+                        $errors[] = gT("You do not have permission on this survey");
+                        continue;
+                    }
                     $oQuota->active = ($action == 'activate' ? 1 : 0);
                     $oQuota->save();
                 } elseif ($action == 'delete') {
+                    if (!Permission::model()->hasSurveyPermission($oQuota->sid, 'quotas', 'delete')) {
+                        $errors[] = gT("You do not have permission on this survey");
+                        continue;
+                    }
                     $oQuota->delete();
                 } elseif ($action == 'changeLanguageSettings' && !empty($_POST['QuotaLanguageSetting'])) {
+                    if (!Permission::model()->hasSurveyPermission($oQuota->sid, 'quotas', 'update')) {
+                        $errors[] = gT("You do not have permission on this survey");
+                        continue;
+                    }
                     $oQuotaLanguageSettings = $oQuota->languagesettings;
                     foreach ($_POST['QuotaLanguageSetting'] as $language => $aQuotaLanguageSettingAttributes) {
                         $oQuotaLanguageSetting = $oQuota->languagesettings[$language];
@@ -271,13 +297,17 @@ class Quotas extends SurveyCommonAction
      */
     public function insertquotaanswer(int $iSurveyId)
     {
+        $this->requirePostRequest();
+
         $iSurveyId = sanitize_int($iSurveyId);
-        $this->checkPermissions($iSurveyId, 'update');
+        $qid = Yii::app()->request->getPost('quota_qid');
+        $quota_id = Yii::app()->request->getPost('quota_id');
+        $this->checkPermissions($iSurveyId, 'update', $quota_id);
 
         $oQuotaMembers = new QuotaMember('create'); // Trigger the 'create' rules
         $oQuotaMembers->sid = $iSurveyId;
-        $oQuotaMembers->qid = Yii::app()->request->getPost('quota_qid');
-        $oQuotaMembers->quota_id = Yii::app()->request->getPost('quota_id');
+        $oQuotaMembers->qid = $qid;
+        $oQuotaMembers->quota_id = $quota_id;
         $oQuotaMembers->code = Yii::app()->request->getPost('quota_anscode');
         if ($oQuotaMembers->save()) {
             if (!empty($_POST['createanother'])) {
@@ -303,15 +333,19 @@ class Quotas extends SurveyCommonAction
      */
     public function delans(int $iSurveyId)
     {
+        $this->requirePostRequest();
+
         $iSurveyId = sanitize_int($iSurveyId);
-        $this->checkPermissions($iSurveyId, 'update');
-
-        QuotaMember::model()->deleteAllByAttributes(array(
-            'id' => Yii::app()->request->getPost('quota_member_id'),
-            'qid' => Yii::app()->request->getPost('quota_qid'),
-            'code' => Yii::app()->request->getPost('quota_anscode'),
-        ));
-
+        $id = App()->request->getPost('quota_member_id');
+        $qid = App()->request->getPost('quota_qid');
+        $code = App()->request->getPost('quota_anscode');
+        /* find related quota by quotamember->id */
+        $QuotaMember = QuotaMember::model()->findByPk($id);
+        if (empty($QuotaMember)) {
+            throw new CHttpException(404, gT("Quota memeber not found."));
+        }
+        $this->checkPermissions($iSurveyId, 'update', $QuotaMember->quota_id);
+        $QuotaMember->delete();
         self::redirectToIndex($iSurveyId);
     }
 
@@ -324,9 +358,8 @@ class Quotas extends SurveyCommonAction
         $this->requirePostRequest();
 
         $iSurveyId = sanitize_int($iSurveyId);
-        $this->checkPermissions($iSurveyId, 'delete');
-
         $quotaId = Yii::app()->request->getQuery('quota_id');
+        $this->checkPermissions($iSurveyId, 'delete', $quotaId);
 
         Quota::model()->deleteByPk($quotaId);
         QuotaLanguageSetting::model()->deleteAllByAttributes(array('quotals_quota_id' => $quotaId));
@@ -343,18 +376,23 @@ class Quotas extends SurveyCommonAction
      */
     public function editquota(int $iSurveyId)
     {
+
         $iSurveyId = sanitize_int($iSurveyId);
         $oSurvey = Survey::model()->findByPk($iSurveyId);
-        $this->checkPermissions($iSurveyId, 'update');
+        $quotaId = Yii::app()->request->getQuery('quota_id');
+
+        $this->checkPermissions($iSurveyId, 'update', $quotaId);
         $aData = $this->getData($iSurveyId);
         $aViewUrls = array();
-        $quotaId = Yii::app()->request->getQuery('quota_id');
+
 
         /* @var Quota $oQuota */
         $oQuota = Quota::model()->findByPk($quotaId);
 
-        if (isset($_POST['Quota'])) {
-            $oQuota->attributes = $_POST['Quota'];
+        if (App()->getRequest()->getPost('Quota')) {
+            $attributes = (array) App()->getRequest()->getPost('Quota');
+            unset($attributes['id']);
+            $oQuota->attributes = $attributes;
             if ($oQuota->save()) {
                 foreach ($_POST['QuotaLanguageSetting'] as $language => $settingAttributes) {
                     $oQuotaLanguageSetting = $oQuota->languagesettings[$language];
@@ -406,13 +444,15 @@ class Quotas extends SurveyCommonAction
      */
     public function newanswer($iSurveyId, $sSubAction = 'newanswer')
     {
+        $this->requirePostRequest();
+
         $iSurveyId = sanitize_int($iSurveyId);
         $oSurvey = Survey::model()->findByPk($iSurveyId);
-
-        $this->checkPermissions($iSurveyId, 'update');
+        $quota_id = Yii::app()->request->getPost('quota_id');
+        $this->checkPermissions($iSurveyId, 'update', $quota_id);
         $aData = $this->getData($iSurveyId);
         $aViewUrls = array();
-        $quota = Quota::model()->findByPk(Yii::app()->request->getPost('quota_id'));
+        $quota = Quota::model()->findByPk($quota_id);
         $aData['oQuota'] = $quota;
 
         if (($sSubAction == "newanswer" || ($sSubAction == "new_answer_two" && !isset($_POST['quota_qid']))) && Permission::model()->hasSurveyPermission($iSurveyId, 'quotas', 'create')) {
@@ -477,8 +517,10 @@ class Quotas extends SurveyCommonAction
         $oQuota = new Quota();
         $oQuota->sid = $oSurvey->primaryKey;
 
-        if (isset($_POST['Quota'])) {
-            $oQuota->attributes = $_POST['Quota'];
+        if (App()->getRequest()->getPost('Quota')) {
+            $aAttributes = (array) App()->getRequest()->getPost('Quota');
+            unset($aAttributes['id']);
+            $oQuota->attributes = $aAttributes;
             if ($oQuota->save()) {
                 foreach ($_POST['QuotaLanguageSetting'] as $language => $settingAttributes) {
                     $oQuotaLanguageSetting = new QuotaLanguageSetting();
